Fixed Issue Where Builds Would Fail
Moved all UnityEdior and System.Linq namespace uses, from the RuleTileFromTemplate.cs script to the RuleTileFromTemplateEditor.cs script. The UnityEditor namespace can't be used in scripts that need to be bundled for build.

TO UPDATE: Delete the old RuleTileFromTemplate.cs and RuleTileFromTemplateEditor.cs files and replace them with the updated versions. If you see errors, you have to click the 'Use Sprite Sheet with Template rules' button again on every 'Rule Tile From Template' asset to re-sort the sprites.